Move all methods handling new credentials to their own module, mask all credentials
