+++ mountpoint-s3/tests/fuse_tests/fork_test.rs
@@ -2,18 +2,18 @@
 #![cfg(feature = "s3_tests")]
 
 use assert_cmd::prelude::*;
-#[cfg(not(feature = "s3express_tests"))]
-use aws_sdk_sts::config::Region;
 use std::fs;
 use std::io::{BufRead, BufReader};
 use std::process::Stdio;
 use std::{path::PathBuf, process::Command};
 use test_case::test_case;
 
-use crate::common::fuse::read_dir_to_entry_names;
-use crate::common::s3::{create_objects, get_test_bucket_and_prefix, get_test_bucket_forbidden, get_test_region};
 #[cfg(not(feature = "s3express_tests"))]
-use crate::common::s3::{get_subsession_iam_role, tokio_block_on};
+use crate::common::creds;
+use crate::common::fuse::read_dir_to_entry_names;
+#[cfg(not(feature = "s3express_tests"))]
+use crate::common::s3::tokio_block_on;
+use crate::common::s3::{create_objects, get_test_bucket_and_prefix, get_test_bucket_forbidden, get_test_region};
 
 const MAX_WAIT_DURATION: std::time::Duration = std::time::Duration::from_secs(10);
 
@@ -396,7 +396,6 @@ fn mount_scoped_credentials() -> Result<(), Box<dyn std::error::Error>> {
     let subprefix = format!("{prefix}sub/");
     let mount_point = assert_fs::TempDir::new()?;
     let region = get_test_region();
-    let subsession_role = get_subsession_iam_role();
 
     // Get scoped down credentials to the subprefix
     let policy = r#"{"Statement": [
@@ -404,18 +403,7 @@ fn mount_scoped_credentials() -> Result<(), Box<dyn std::error::Error>> {
         {"Effect": "Allow", "Action": "s3:ListBucket", "Resource": "arn:aws:s3:::__BUCKET__", "Condition": {"StringLike": {"s3:prefix": "__PREFIX__*"}}}
     ]}"#;
     let policy = policy.replace("__BUCKET__", &bucket).replace("__PREFIX__", &subprefix);
-    let config = tokio_block_on(aws_config::from_env().region(Region::new(get_test_region())).load());
-    let sts_client = aws_sdk_sts::Client::new(&config);
-    let credentials = tokio_block_on(
-        sts_client
-            .assume_role()
-            .role_arn(subsession_role)
-            .role_session_name("test_scoped_credentials")
-            .policy(policy)
-            .send(),
-    )
-    .unwrap();
-    let credentials = credentials.credentials().unwrap();
+    let credentials = tokio_block_on(creds::get_scoped_down_credentials(policy));
 
     // First try without the subprefix -- mount should fail as we don't have permissions on it
     let mut cmd = Command::cargo_bin("mount-s3")?;
@@ -425,9 +413,14 @@ fn mount_scoped_credentials() -> Result<(), Box<dyn std::error::Error>> {
         .arg(format!("--prefix={prefix}"))
         .arg("--auto-unmount")
         .arg(format!("--region={region}"))
-        .env("AWS_ACCESS_KEY_ID", credentials.access_key_id().unwrap())
-        .env("AWS_SECRET_ACCESS_KEY", credentials.secret_access_key().unwrap())
-        .env("AWS_SESSION_TOKEN", credentials.session_token().unwrap())
+        .env("AWS_ACCESS_KEY_ID", credentials.access_key_id())
+        .env("AWS_SECRET_ACCESS_KEY", credentials.secret_access_key())
+        .env(
+            "AWS_SESSION_TOKEN",
+            credentials
+                .session_token()
+                .expect("session credentials should always have session token"),
+        )
         .spawn()
         .expect("unable to spawn child");
 
@@ -454,9 +447,14 @@ fn mount_scoped_credentials() -> Result<(), Box<dyn std::error::Error>> {
         .arg(format!("--prefix={subprefix}"))
         .arg("--auto-unmount")
         .arg(format!("--region={region}"))
-        .env("AWS_ACCESS_KEY_ID", credentials.access_key_id().unwrap())
-        .env("AWS_SECRET_ACCESS_KEY", credentials.secret_access_key().unwrap())
-        .env("AWS_SESSION_TOKEN", credentials.session_token().unwrap())
+        .env("AWS_ACCESS_KEY_ID", credentials.access_key_id())
+        .env("AWS_SECRET_ACCESS_KEY", credentials.secret_access_key())
+        .env(
+            "AWS_SESSION_TOKEN",
+            credentials
+                .session_token()
+                .expect("session credentials should always have session token"),
+        )
         .spawn()
         .expect("unable to spawn child");
